requirements.txt
Flask==2.3.2
requests==2.32.3
python-dateutil==2.9.0.post0
ipeadatapy==0.1.1
gunicorn==20.1.0